specify crosshair size in pixels

+++ include/Hyper/Game.hpp
@@ -42,7 +42,7 @@ namespace Game {
     }
 
     namespace GUI {
-        extern GLfloat aimSize;
+        extern int aimSize;
     }
 
     namespace Keyboard {
